requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
discord==1.7.3
discord-py-interactions==4.0.1
discord-py-slash-command==3.0.1
discord.py==1.7.3
psycopg2-binary==2.9.1